added beginning coin return tests

+++ PillarKata.VendingMachine/PillarKata.VendingMachine/Coin.cs
@@ -2,9 +2,9 @@ namespace PillarKata.VendingMachine
 {
     public class Coin
     {
-        public int WeightInGrams { get; private set; }
+        public double WeightInGrams { get; private set; }
 
-        public Coin(int weightInGrams)
+        public Coin(double weightInGrams)
         {
             WeightInGrams = weightInGrams;
         }
